update CI job
