+++ playwright/package.json
@@ -17,6 +17,7 @@
     "@types/node": "^20.5.0",
     "@typescript-eslint/eslint-plugin": "^5.59.10",
     "@typescript-eslint/parser": "^5.59.10",
+    "axios": "^1.4.0",
     "babel-eslint": "^10.1.0",
     "dotenv": "^8.6.0",
     "eslint": "^8.42.0",
@@ -30,6 +31,6 @@
     "timediff": "^1.1.1",
     "tyk-test-automation-wrappers": "git+https://github.com/TykTechnologies/tyk-test-automation-wrappers.git#main",
     "typescript": "^5.1.3",
-    "uuidv4": "^6.2.12"
+    "uuid": "^9.0.0"
   }
 }
