+++ playwright/tests/specs/policies/policy_settings_test.ts
@@ -11,7 +11,7 @@ import { test, assert } from '@fixtures';
 
 import { policies_page } from '../../../lib/pom/Policies_page';
 import { Dashboard_connection } from '@api_connections/Dashboard_connection';
-import { newAPIdefinitionWithDefaults } from '../../../lib/utils/API_object_designer';
+import { newAPIdefinitionWithDefaults } from '@lib/utils/API_object_designer';
 
 const testApiDetails = {
   apiName: 'test_api'
@@ -47,7 +47,7 @@ test('Create policy page - settings tests', async ({ createUserAndLogin, main_pa
 
   await test.step('Prerequisits: creating API definition via dashboard API', async () => {
     const apiDefinition = newAPIdefinitionWithDefaults({"name" : "test_api"});
-    dashboard_connection.createAPI(apiDefinition, envDetails.userSecret);
+    await dashboard_connection.createAPI(apiDefinition, envDetails.userSecret);
   });
 
   await test.step('User should be able to create policy with Access Denied status', async () => {
