+++ playwright/tests/specs/policies/policy_tags_test.ts
@@ -2,7 +2,7 @@ import { test, assert } from '@fixtures';
 
 import { policies_page } from '../../../lib/pom/Policies_page';
 import { Dashboard_connection } from '@api_connections/Dashboard_connection';
-import { newAPIdefinitionWithDefaults } from '../../../lib/utils/API_object_designer';
+import { newAPIdefinitionWithDefaults } from '@lib/utils/API_object_designer';
 
 const policyDetails = {
   apiName: 'test_api',
@@ -25,7 +25,7 @@ test('Create/update/delete tags on policy', async ({ createUserAndLogin, main_pa
 
   await test.step('Prerequisits: creating API definition via dashboard API', async () => {
     const apiDefinition = newAPIdefinitionWithDefaults({"name":policyDetails.apiName});
-    dashboard_connection.createAPI(apiDefinition, envDetails.userSecret);
+    await dashboard_connection.createAPI(apiDefinition, envDetails.userSecret);
   });
 
   await test.step('User should be able to create new Policy with tag', async () => {
