+++ playwright/tsconfig.json
@@ -31,6 +31,7 @@
     "baseUrl": "./",                                  /* Specify the base directory to resolve non-relative module names. */
     "paths": {
       "@utils":["lib/utils.ts"],
+      "@lib/*":["lib/*"],
       "@api_connections/*":["lib/utils/api_connections/*"],
       "@pom/*":["lib/pom/*"],
       "@variables":["config_variables.ts"],
